refactor and integrate with insertScreen
