+++ main.c
@@ -1,7 +1,5 @@
 #include "chestLib.h"
-#include "interface/interfaceLib.h"
-#include <stdio.h>
-#include <stdlib.h>
+#include "chestLib.h"
 
 int opt = 0, page = HOME;
 telaInserir inserir;
@@ -10,12 +8,15 @@ int lista[31];
 listagemEmTela listagem;
 
 int main(int argc, const char * argv[]) {
-    Chest * chest = (Chest *) malloc(sizeof(Chest));
-    //initializeChest(chest);
-    //printf("%d\n", hashFunction(256, "PICKAXE"));
+    chest = createChest();
+
+    int pontIndice = 0;
+
+	initializeHashCodes();
+	initializeItensLib();
 
 	char tecla = 0;
-	createListaObjetos();
+
 	int antPage = page;
 	system("MODE 140, 40");
 	system("title Minecraft");
@@ -28,7 +29,7 @@ int main(int argc, const char * argv[]) {
 			showHome(tecla);
 			break;
 		case INSERIR:
-			showInserir(tecla);
+			showInserir(tecla, &pontIndice);
 			break;
 		case REMOVER:
 			showRemover(tecla);
