+++ showInserir.c
@@ -1,8 +1,8 @@
-#include "interfaceLib.h"
+#include "chestLib.h"
+#include "chestLib.h"
 
-void showInserir(int tecla)
-{
-int itemsPerPage = 7, altura = 2;
+void showInserir(int tecla, int *pontIndice) {
+	int itemsPerPage = 7, altura = 2;
 
 	if (tecla == 119 || tecla == 87){ // caso w|W for pressionado
 		inserir.opt--;
@@ -12,8 +12,7 @@ int itemsPerPage = 7, altura = 2;
 		if (inserir.opt < inserir.page * itemsPerPage){
 				inserir.page--;
 		}
-	}
-		if (tecla == 115 || tecla == 83){ // caso s|S for pressionado
+	} if (tecla == 115 || tecla == 83){ // caso s|S for pressionado
 			inserir.opt++;
 			if(inserir.opt >30) inserir.opt=30;
 
@@ -75,7 +74,7 @@ int itemsPerPage = 7, altura = 2;
 			moldura(LARGURA/2, LARGURA-10, ALTURA/2 + y, ALTURA/2 +altura + y);
 			gotoxy(LARGURA/2 + 2,ALTURA/2 + altura/2 + y );
 			white();
-			printf("%d", lista[inserir.page*itemsPerPage+i]);
+			printf("%s", itensMineCraft[inserir.page*itemsPerPage+i]->name);
 			y += altura+1;
 		}
 
@@ -94,31 +93,62 @@ int itemsPerPage = 7, altura = 2;
 		printf (">>");
 	}
 
-	if(inserir.escolheuItem == 1){
+	if(inserir.escolheuItem > 0){
 		gotoxy((LARGURA - 30)/2, 9);
 		printf ("QUANTIDADE: ");
 		gotoxy((LARGURA)/2, 9);
 		scanf("%d", &inserir.qtd);
-		inserir.escolheuItem++;
-	}
 
-	if(inserir.escolheuItem == 2){
-		if(inserir.qtd > 64){
+		if(tecla == 109 || tecla == 77){ // caso m|M for pressionado
+            page = HOME;
+        }
+		// seria uma boa verificar se qtd e valida aqui
+
+		Item* item = itensMineCraft[inserir.opt];
+		item->quantity = inserir.qtd;
+
+
+		if(hashCodes[inserir.opt] == 0) {
+			hashCodes[inserir.opt] = hashFunction(item->key, item->name);
+		}
+
+		int response = insertNode(item, hashCodes[inserir.opt]);
+
+		if(response == 0) {
+			desenhaAviso("O bau esta cheio.");
+			Sleep(2000);
+			inserir.escolheuItem = 1;
+		}
+
+		if(response == 1) {
+			desenhaAviso("Insira um numero valido!");
+			Sleep(2000);
+			inserir.escolheuItem = 1;
+		}
+
+		if(response == 2  || response == 4) {
 			desenhaAviso("Quantidade excede o limite.");
 			Sleep(2000);
 			inserir.escolheuItem = 1;
-		}else if(inserir.qtd < 0){
-			desenhaAviso("Insira um numero válido!");
-			Sleep(2000);
-			inserir.escolheuItem = 1;
-		}else{
+		}
+
+		if(response == 3 || response == 5){
 			desenhaAviso("Salvo com sucesso!");
+			page = HOME;
+            Sleep(2000);
+
+		}
+
+		if(response == 6) {
+			desenhaAviso("Eita! Houve colisao, mas foi salvo com sucesso!");
 			Sleep(2000);
 			page = HOME;
 		}
 
+		item->quantity = inserir.qtd;
 
-
+		inserir.escolheuItem++;
 	}
+
 	desenhaBauAberto(20, 15);
 }
