+++ showSearch.c
@@ -1,4 +1,4 @@
-#include "interfaceLib.h"
+#include "chestLib.h"
 
 void showBuscar(int tecla)
 {
